made animation FPS

+++ source/_struct.h
@@ -475,7 +475,7 @@ typedef struct
 	vec3_t Angle;
 	vec3_t Light;
 	float  Alpha;
-	int    LifeTime;
+	float    LifeTime;
 	OBJECT *Target;
 	float  Rotation;
 	int    Frame;
